requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
gunicorn
psycopg2-binary